Each library system moved into a separate file
